[chore] 웨이블존 관련 Admin Repository 로직을 분리
