+++ src/main/java/com/wayble/server/admin/repository/AdminWaybleZoneRepository.java
@@ -1,0 +1,7 @@
+package com.wayble.server.admin.repository;
+
+import com.wayble.server.wayblezone.entity.WaybleZone;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AdminWaybleZoneRepository extends JpaRepository<WaybleZone, Long>, AdminWaybleZoneRepositoryCustom {
+}
